Rename userCard and renderTemplates file
