+++ lib/commands/showStats.js
@@ -3,7 +3,7 @@ const propertyToKey = require('../adapters/propertyToKey')
 const mapCycleTimeToStory = require('../adapters/mapCycleTimeToStory')
 const memberIdToKey = require('../adapters/memberIdToKey')
 const mapStoryToMember = require('../adapters/mapStoryToMember')
-const renderTemplates = require('../stats/renderTemplates')
+const getAllMemberCards = require('../stats/getAllMemberCards')
 const createDoc = require('../stats/createDoc')
 const statsCss = require('../stats/css/statsStyle')
 const {window, ViewColumn} = require('vscode')
@@ -19,7 +19,7 @@ module.exports = async context => {
   mapCycleTimeToStory(allIterationStories, iterationCycle.data)
   mapStoryToMember(allMembersById, allIterationStories)
 
-  const dataView = renderTemplates(allMembersById)
+  const dataView = getAllMemberCards(allMembersById)
   const statsHtml = createDoc(dataView, statsCss)
   const statsPanel = window.createWebviewPanel('pivotaly.stats', 'Statistics', ViewColumn.One, {enableFindWidget: true})
   statsPanel.webview.html = statsHtml
