+++ lib/stats/getAllMemberCards.js
@@ -1,5 +1,5 @@
 const _ = require('lodash')
-const renderMemberCard = require('./templates/userCard')
+const renderMemberCard = require('./templates/prepareUserCard')
 
 module.exports = members => {
   let renderedData = ''
