+++ lib/stats/templates/prepareUserCard.js
